# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TourVoice API",
    "description": "Hệ thống Thuyết minh Du lịch Tự động Đa ngôn ngữ Quận 4 (FastAPI + PyMongo Async + MongoDB).",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/pois": {
      "get": {
        "tags": [
          "POIs (Points of Interest)"
        ],
        "summary": "List Public Pois",
        "description": "Lists public POIs with language fallback.",
        "operationId": "list_public_pois_api_v1_pois_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Requested language code (vi, en, fr, zh, ja)",
              "default": "vi",
              "title": "Lang"
            },
            "description": "Requested language code (vi, en, fr, zh, ja)"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Public Pois Api V1 Pois Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "POIs (Points of Interest)"
        ],
        "summary": "Create Poi Admin",
        "description": "Creates a new POI (Admin only).",
        "operationId": "create_poi_admin_api_v1_pois_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/POICreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/pois/{poi_id}/menu": {
      "get": {
        "tags": [
          "Menu Items"
        ],
        "summary": "List Poi Menu",
        "description": "Use Case T06: Public menu view for a POI.",
        "operationId": "list_poi_menu_api_v1_pois__poi_id__menu_get",
        "parameters": [
          {
            "name": "poi_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Poi Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Poi Menu Api V1 Pois  Poi Id  Menu Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Menu Items"
        ],
        "summary": "Create Menu Item",
        "description": "Use Case C05, O09: Add a dish to POI menu.",
        "operationId": "create_menu_item_api_v1_pois__poi_id__menu_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "poi_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Poi Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuItemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/menu/{item_id}": {
      "patch": {
        "tags": [
          "Menu Items"
        ],
        "summary": "Update Menu Item",
        "description": "Use Case C05, O09: Update dish.",
        "operationId": "update_menu_item_api_v1_menu__item_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Menu Items"
        ],
        "summary": "Delete Menu Item",
        "description": "Use Case C05, O09: Delete dish or deactivate.",
        "operationId": "delete_menu_item_api_v1_menu__item_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/owner/registration": {
      "get": {
        "tags": [
          "Store Owner Portal (Chủ quán)"
        ],
        "summary": "Get Registration Status",
        "description": "Use Case O02: Check owner verification status.",
        "operationId": "get_registration_status_api_v1_owner_registration_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/owner/pois": {
      "get": {
        "tags": [
          "Store Owner Portal (Chủ quán)"
        ],
        "summary": "Get My Pois",
        "description": "Use Case O03: View list of POIs owned by current owner.",
        "operationId": "get_my_pois_api_v1_owner_pois_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/owner/submissions": {
      "get": {
        "tags": [
          "Store Owner Portal (Chủ quán)"
        ],
        "summary": "List My Submissions",
        "description": "Use Case O06: View owner's submissions and admin feedback.",
        "operationId": "list_my_submissions_api_v1_owner_submissions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Store Owner Portal (Chủ quán)"
        ],
        "summary": "Create Submission",
        "description": "Use Case O04, O05: Submit draft POI or modification for Admin review.",
        "operationId": "create_submission_api_v1_owner_submissions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/owner/notifications": {
      "get": {
        "tags": [
          "Store Owner Portal (Chủ quán)"
        ],
        "summary": "List My Notifications",
        "description": "Use Case O07: View in-app notifications.",
        "operationId": "list_my_notifications_api_v1_owner_notifications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/owner/notifications/{notif_id}/read": {
      "patch": {
        "tags": [
          "Store Owner Portal (Chủ quán)"
        ],
        "summary": "Mark Notification As Read",
        "description": "Use Case O08: Mark notification as read.",
        "operationId": "mark_notification_as_read_api_v1_owner_notifications__notif_id__read_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notif_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notif Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/owner/analytics": {
      "get": {
        "tags": [
          "Store Owner Portal (Chủ quán)"
        ],
        "summary": "Get Owner Analytics",
        "description": "Use Case O10: View listening statistics for owned POIs.",
        "operationId": "get_owner_analytics_api_v1_owner_analytics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GeoLocation": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "Point"
          },
          "coordinates": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Coordinates",
            "description": "[longitude, latitude]"
          }
        },
        "type": "object",
        "required": [
          "coordinates"
        ],
        "title": "GeoLocation"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MenuItemCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "default": ""
          },
          "price": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Price",
            "description": "Price in lowest integer unit (VND)"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "VND"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "price"
        ],
        "title": "MenuItemCreate"
      },
      "MenuItemUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          }
        },
        "type": "object",
        "title": "MenuItemUpdate"
      },
      "POICreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "minLength": 5,
            "title": "Description"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "sightseeing"
          },
          "address": {
            "type": "string",
            "title": "Address",
            "default": ""
          },
          "location": {
            "$ref": "#/components/schemas/GeoLocation"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Images"
          },
          "trigger_radius": {
            "type": "number",
            "maximum": 500.0,
            "minimum": 5.0,
            "title": "Trigger Radius",
            "default": 30.0
          },
          "audio_priority": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Audio Priority",
            "default": 1
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "source_lang": {
            "type": "string",
            "title": "Source Lang",
            "default": "vi"
          },
          "activation_requested": {
            "type": "boolean",
            "title": "Activation Requested",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "location"
        ],
        "title": "POICreate"
      },
...
      "SubmissionCreateRequest": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action",
            "description": "'create' or 'update'",
            "default": "create"
          },
          "poi_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poi Id"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "request_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Key"
          }
        },
        "type": "object",
        "required": [
          "payload"
        ],
        "title": "SubmissionCreateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/admin/auth/login"
          }
        }
      }
    }
  }
}